This is what was output by my work emacs. See if this works at home.
I'm really quite confused by how they handle fonts in emacs. For some
reason I had to switch off of my Iosevka font and back on to it before
it would actually change into the Iosevka font. Something was really
broken because the *describe-font* function was saying I was using
Iosevka15 but I know it wasn't that at all. I believe it was just the
default Ubuntu font. Anyways, after changing it off and back on again,
I'm finally happy that it's using the font I wanted.